backend/requirements.txt
Django>=6.0a1,<7.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
# mysqlclient>=2.2
django-filter>=24.1
nanoid>=2.0
google-auth>=2.29
requests
python-dotenv